feat: introduce and use a new feature flag for flexible layouts
The currently used flexible layout feature flag
enableDashboardFlexibleLayout is not on the dashboard “unfinished
features” blacklist. When it was enabled, the unfinished version of
the layouts was enabled in customer apps that use old SDKs.

To resolve the situation, we will create a new feature flag that
will be used instead of it: enableFlexibleDashboardLayout. It will
also replace enableDashboardFlexibleLayoutContainer, the feature flag
that we no longer need.

JIRA: LX-1339
risk: low